Add files via upload
CNTFET-Based Ternary Full-Adder Design using Unary Operators

The energy consumption has to be reduced for designing  the embedded systems . Through Multiple-valued logic (MVL) circuits, interconnection complexity and energy consumption their is decrease in comparison to binary systems. In this paper we used MVL circuits to present a ternary full-adder (TFA) which reduces the consumption of the energy for the maintenance of the battery  in nano-scale embedded systems and IoT devices. We used CNTFET-based circuit which uses a dual- voltage (Vdd and Vdd/2) and novel unary operators to improve the performance of the designed system.